Add sketch TinyAlpacaServerDemo, compiles on host... with quite a few stubbed out methods still to write.
Flesh out more of DeviceApiHandlerBase and ObservingConditionsAdapter.

Add ErrorCode members for all the standard ASCOM errors.

+++ src/utils/escaping.cpp
@@ -15,7 +15,7 @@ size_t PrintJsonEscaped::write(uint8_t b) {
 
 size_t PrintJsonEscaped::write(const uint8_t* buffer, size_t size) {
   size_t count = 0;
-  for (int ndx = 0; ndx < size; ++ndx) {
+  for (size_t ndx = 0; ndx < size; ++ndx) {
     count += PrintCharJsonEscaped(wrapped_, static_cast<char>(buffer[ndx]));
   }
   return count;
